requirements.txt
fastapi
uvicorn[standard]
pypdf
pdfplumber
python-docx
spacy
supabase
python-dotenv
google-genai
google-auth
pydantic
pydantic-settings
python-multipart
